Made the sample more useful with providing colors in extra functions. Add a new function to create bitmap image for square data fast, but with unsafe code. Renamed the sample, so that it is more a sample for an image over the map, but not a weather app.
